Added revert.sh
Used in lab environment to revert back to pristine snapshot state.
Only one snapshot is made of the VMs.  If multiple snapshots exist, not
sure what will happen.

This is running a query of all VMs with the name "mapr".

Executed on ESXi host directly or via ssh root@hostname < revert.sh